# Optimize formatter performance with StringBuilder mod