--- a/internal/pool.go
+++ b/internal/pool.go
@@ -1,7 +1,6 @@
 package internal
 
 import (
-	"bytes"
 	"sync"
 )
 
@@ -10,18 +9,18 @@ var buffers *sync.Pool
 func init() {
 	buffers = &sync.Pool{
 		New: func() interface{} {
-			return new(bytes.Buffer)
+			return new(StringBuilder)
 		},
 	}
 }
 
 // GetBuffer fetches a buffers from the pool
-func GetBuffer() *bytes.Buffer {
-	return buffers.Get().(*bytes.Buffer)
+func GetBuffer() *StringBuilder {
+	return buffers.Get().(*StringBuilder)
 }
 
 // PutBuffer returns a buffers to the pool
-func PutBuffer(buf *bytes.Buffer) {
+func PutBuffer(buf *StringBuilder) {
 	buf.Reset()
 	buffers.Put(buf)
 }