… internal/string_builder.go