The Go standard StringBuilder has some limitations, this pr included a modified version of StringBuilder. The changes include: 

- do not free the underlying buffer when Reset(); so it can be reused in pool
- add AppendFloat and AppendInt methods; to avoid additional alloc caused by strconv.Format*
- add Bytes() method

Besides the mod above also include:

- cleanup unnecessary conditions in sanitize
- remove the additional allocations used by sanitize

## Benchmark Result

cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz

Original:

```
BenchmarkMetricLine-12 875883 1229 ns/op 250 B/op 17 allocs/op
BenchmarkHistoLine-12 635227 1729 ns/op 614 B/op 22 allocs/op
BenchmarkSpanLine-12 1623225 747.5 ns/op 340 B/op 11 allocs/op
```

Optimized:

```
BenchmarkMetricLine-12 3926373 309.9 ns/op 64 B/op 1 allocs/op
BenchmarkHistoLine-12 1764009 674.7 ns/op 416 B/op 6 allocs/op
BenchmarkSpanLine-12 4476735 266.7 ns/op 192 B/op 1 allocs/op
```

Summary:

| Method     | CPU Speedup | Alloc Reduction |
|------------|-------------|-----------------|
| MetricLine | 4.0x        | -94%            |
| HistoLine  | 2.5x        | -72%            |
| SpanLine   | 2.8x        | -91%            |