@fredwangwang 2021-07-02
failure on testing against go 1.11 for missing `strings.ReplaceAll`. Is that version (almost 3 years ago) still need to be supported?