@keep94 2022-10-21
Closing this PR.  Since this is an API that makes network calls, most of the time will be spent doing HTTP requests. It seems like tweaking StringBuilder would be premature optimization.  However, if you are able to justify the need for this optimization as well as provide test coverage for the modified StringBuilder, we would be happy to review this once again. Thank you for contributing.